requirements.txt
-e git+https://github.com/cannibalcheeseburger/automation-cli.git@4dcd14088108c3d15c7a3376bf927711b1f97f18#egg=automate
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
idna==2.9
IMDbPY==6.8
lxml==4.5.0
requests==2.23.0
selenium==3.141.0
SQLAlchemy==1.3.17
urllib3==1.25.9
